backend/requirements.txt
Flask>=2.3.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
